+++ core/fslibs/parsing_routines/__init__.py
@@ -4,4 +4,4 @@ from core.fslibs.parsing_routines.ansig import parse_ansig_peaklist as ansig
 from core.fslibs.parsing_routines.ccpnmrv2 import parse_ccpnmrv2_peaklist as ccpnmrv2
 #from core.fslibs.parsing_routines.nmrdraw import nmrdraw
 #from core.fslibs.parsing_routines.nmrview import nmrview
-#from core.fslibs.parsing_routines.sparky import sparky
+from core.fslibs.parsing_routines.sparky import parse_sparky_peaklist as sparky
